Create base directory and move top level under it
