+++ client/base/configureStore.js
