+++ client/base/index.js
@@ -1,0 +1,2 @@
+export Router from './Router'
+export configureStore from './configureStore'
